requirements.txt
streamlit
Pillow
pytesseract
PyMuPDF
google-generativeai
